refactor: address review findings for Stripe subscription API
- Move webhook business logic from route handler to subscriptionService
- Use domain errors (NotFoundError, InvalidStateError) instead of plain Error
- Fix import order: group @nexus/db/* before @/ imports
- Simplify JSDoc on stripe namespace to explain WHY not WHAT
- Remove unnecessary `as PlanTier` cast
- Guard against status regression in handlePaymentFailed
- Add subscriptions to mock DB for future unit tests
- Replace DEFAULT_STORAGE_LIMIT_BYTES with PLAN_LIMITS.starter
- Use shared PlanTier type in storageService
- Derive Zod enums from shared CHECKOUT_TIERS/BILLING_INTERVALS consts
