+++ apps/web/app/api/webhooks/stripe/route.ts
@@ -1,52 +1,12 @@
 import { NextResponse, type NextRequest } from 'next/server';
 import type Stripe from 'stripe';
 import { createWebhookRepo } from '@nexus/db/repo/webhooks';
-import { createSubscriptionRepo } from '@nexus/db/repo/subscriptions';
 import { db } from '@/server/db';
 import { logger } from '@/server/lib/logger';
 import { stripe } from '@/lib/stripe';
-import { PLAN_LIMITS, type PlanTier } from '@/server/services/constants';
-import { subscriptionStatusEnum } from '@nexus/db/schema';
+import { subscriptionService } from '@/server/services/subscriptions';
 
 const log = logger.child({ handler: 'stripe-webhook' });
-
-const VALID_STATUSES = new Set<string>(subscriptionStatusEnum.enumValues);
-
-/** Stripe uses expandable fields (string | object | null) — normalize to the ID string. */
-function resolveStripeId(
-    field: string | { id: string } | null | undefined
-): string | undefined {
-    if (!field) return undefined;
-    return typeof field === 'string' ? field : field.id;
-}
-
-type SubscriptionStatus = (typeof subscriptionStatusEnum.enumValues)[number];
-
-/** Map Stripe status to our DB enum, falling back to existing value for unknown statuses. */
-function mapStripeStatus(
-    stripeStatus: string,
-    fallback: SubscriptionStatus
-): SubscriptionStatus {
-    if (VALID_STATUSES.has(stripeStatus)) {
-        return stripeStatus as SubscriptionStatus;
-    }
-    log.warn(
-        { stripeStatus },
-        'Unknown Stripe subscription status, using fallback'
-    );
-    return fallback;
-}
-
-function resolveTierFromSubscription(
-    sub: Stripe.Subscription
-): PlanTier | null {
-    const item = sub.items.data[0];
-    if (!item) return null;
-    const product = item.price.product as Stripe.Product | string;
-    const metadata =
-        typeof product === 'string' ? null : (product.metadata ?? null);
-    return (metadata?.tier as PlanTier) ?? null;
-}
 
 export async function POST(request: NextRequest): Promise<NextResponse> {
     let rawBody: string;
@@ -115,7 +75,7 @@ export async function POST(request: NextRequest): Promise<NextResponse> {
     log.info({ eventId: event.id, eventType: event.type }, 'Webhook received');
 
     try {
-        await dispatch(event);
+        await subscriptionService.dispatchWebhookEvent(db, event);
 
         await webhookRepo.update(webhookEvent.id, { status: 'processed' });
 
@@ -142,162 +102,4 @@ export async function POST(request: NextRequest): Promise<NextResponse> {
 
         return NextResponse.json({ received: true });
     }
-}
-
-async function dispatch(event: Stripe.Event): Promise<void> {
-    switch (event.type) {
-        case 'checkout.session.completed':
-            await handleCheckoutCompleted(
-                event.data.object as Stripe.Checkout.Session
-            );
-            break;
-        case 'customer.subscription.created':
-        case 'customer.subscription.updated':
-            await handleSubscriptionUpsert(
-                event.data.object as Stripe.Subscription
-            );
-            break;
-        case 'customer.subscription.deleted':
-            await handleSubscriptionDeleted(
-                event.data.object as Stripe.Subscription
-            );
-            break;
-        case 'invoice.payment_failed':
-            await handlePaymentFailed(event.data.object as Stripe.Invoice);
-            break;
-        default:
-            log.debug({ eventType: event.type }, 'Unhandled event type');
-    }
-}
-
-async function handleCheckoutCompleted(
-    session: Stripe.Checkout.Session
-): Promise<void> {
-    if (session.mode !== 'subscription') return;
-
-    const customerId = resolveStripeId(session.customer);
-    const subscriptionId = resolveStripeId(session.subscription);
-
-    if (!customerId || !subscriptionId) {
-        log.warn(
-            { sessionId: session.id },
-            'Checkout session missing customer or subscription'
-        );
-        return;
-    }
-
-    const repo = createSubscriptionRepo(db);
-    const sub = await repo.findByStripeCustomerId(customerId);
-    if (!sub) {
-        log.warn(
-            { customerId, sessionId: session.id },
-            'No subscription record for customer'
-        );
-        return;
-    }
-
-    // subscription.created will fill in plan details; this just links the IDs
-    await repo.upsertFromWebhook({
-        ...sub,
-        stripeSubscriptionId: subscriptionId,
-    });
-
-    log.info(
-        { customerId, subscriptionId },
-        'Linked Stripe subscription from checkout'
-    );
-}
-
-async function handleSubscriptionUpsert(
-    sub: Stripe.Subscription
-): Promise<void> {
-    const customerId = resolveStripeId(sub.customer);
-    if (!customerId) return;
-
-    const repo = createSubscriptionRepo(db);
-    const existing = await repo.findByStripeCustomerId(customerId);
-    if (!existing) {
-        log.warn(
-            { customerId, stripeSubscriptionId: sub.id },
-            'No local record for subscription upsert'
-        );
-        return;
-    }
-
-    const tier = resolveTierFromSubscription(sub) ?? existing.planTier;
-    const storageLimit = PLAN_LIMITS[tier as PlanTier] ?? existing.storageLimit;
-
-    // In API version 2026-02-25.clover, period fields moved from Subscription to its items
-    const firstItem = sub.items.data[0];
-    const periodStart = firstItem
-        ? new Date(firstItem.current_period_start * 1000)
-        : existing.currentPeriodStart;
-    const periodEnd = firstItem
-        ? new Date(firstItem.current_period_end * 1000)
-        : existing.currentPeriodEnd;
-
-    await repo.upsertFromWebhook({
-        ...existing,
-        stripeSubscriptionId: sub.id,
-        planTier: tier,
-        status: mapStripeStatus(sub.status, existing.status),
-        storageLimit,
-        currentPeriodStart: periodStart,
-        currentPeriodEnd: periodEnd,
-        cancelAtPeriodEnd: sub.cancel_at_period_end,
-        trialEnd: sub.trial_end ? new Date(sub.trial_end * 1000) : null,
-    });
-
-    log.info(
-        { customerId, tier, status: sub.status },
-        'Subscription synced from webhook'
-    );
-}
-
-async function handleSubscriptionDeleted(
-    sub: Stripe.Subscription
-): Promise<void> {
-    const customerId = resolveStripeId(sub.customer);
-    if (!customerId) return;
-
-    const repo = createSubscriptionRepo(db);
-    const existing = await repo.findByStripeCustomerId(customerId);
-    if (!existing) {
-        log.warn({ customerId }, 'No local record for subscription deletion');
-        return;
-    }
-
-    await repo.upsertFromWebhook({
-        ...existing,
-        status: 'canceled',
-        cancelAtPeriodEnd: false,
-    });
-
-    log.info({ customerId }, 'Subscription marked as canceled');
-}
-
-async function handlePaymentFailed(invoice: Stripe.Invoice): Promise<void> {
-    const customerId = resolveStripeId(invoice.customer);
-
-    if (!customerId) {
-        log.warn({ invoiceId: invoice.id }, 'Invoice missing customer');
-        return;
-    }
-
-    const repo = createSubscriptionRepo(db);
-    const existing = await repo.findByStripeCustomerId(customerId);
-    if (!existing) {
-        log.warn({ customerId }, 'No local record for payment failure');
-        return;
-    }
-
-    await repo.upsertFromWebhook({
-        ...existing,
-        status: 'past_due',
-    });
-
-    log.info(
-        { customerId, invoiceId: invoice.id },
-        'Subscription marked past_due'
-    );
 }
